+++ include/libstd/sys/io.h
@@ -20,6 +20,7 @@ extern "C"
 #include "_internal/clock_t.h"
 #include "_internal/config.h"
 #include "_internal/fd_t.h"
+#include "_internal/ssize_t.h"
 #include "_internal/time_t.h"
 
 /**
@@ -127,7 +128,7 @@ uint64_t close(fd_t fd);
  * @return On success, the number of bytes read. On end-of-file, 0. On failure, `ERR` and `errno`
  * is set.
  */
-uint64_t read(fd_t fd, void* buffer, uint64_t count);
+size_t read(fd_t fd, void* buffer, size_t count);
 
 /**
  * @brief Wrapper for reading a file directly into a null-terminated string.
@@ -151,7 +152,7 @@ char* sread(fd_t fd);
  * @param offset The offset in the file to start reading from.
  * @return On success, the number of bytes read. On end-of-file, 0. On failure, `ERR` and `errno` is set.
  */
-uint64_t readfile(const char* path, void* buffer, uint64_t count, uint64_t offset);
+size_t readfile(const char* path, void* buffer, size_t count, size_t offset);
 
 /**
  * @brief Wrapper for reading an entire file directly into a null-terminated string.
@@ -174,7 +175,7 @@ char* sreadfile(const char* path);
  * @param count The number of bytes to write.
  * @return On success, the number of bytes written. On failure, `ERR` and `errno` is set.
  */
-uint64_t write(fd_t fd, const void* buffer, uint64_t count);
+size_t write(fd_t fd, const void* buffer, size_t count);
 
 /**
  * @brief Wrapper for writing a null-terminated string to a file.
@@ -183,7 +184,7 @@ uint64_t write(fd_t fd, const void* buffer, uint64_t count);
  * @param string The null-terminated string to write.
  * @return On success, the number of bytes written. On failure, `ERR` and `errno` is set.
  */
-uint64_t swrite(fd_t fd, const char* string);
+size_t swrite(fd_t fd, const char* string);
 
 /**
  * @brief Wrapper for writing to a file directly using a path.
@@ -196,7 +197,7 @@ uint64_t swrite(fd_t fd, const char* string);
  * @param offset The offset in the file to start writing to.
  * @return On success, the number of bytes written. On failure, `ERR` and `errno` is set.
  */
-uint64_t writefile(const char* path, const void* buffer, uint64_t count, uint64_t offset);
+size_t writefile(const char* path, const void* buffer, size_t count, size_t offset);
 
 /**
  * @brief Wrapper for writing a null-terminated string directly to a file using a path.
@@ -207,7 +208,7 @@ uint64_t writefile(const char* path, const void* buffer, uint64_t count, uint64_
  * @param string The null-terminated string to write.
  * @return On success, the number of bytes written. On failure, `ERR` and `errno` is set.
  */
-uint64_t swritefile(const char* path, const char* string);
+size_t swritefile(const char* path, const char* string);
 
 /**
  * @brief Wrapper for reading from a file descriptor using scan formatting.
@@ -266,7 +267,7 @@ typedef uint8_t seek_origin_t;
  * @return On success, the new offset from the beginning of the file. On failure, `ERR` and `errno` is
  * set.
  */
-uint64_t seek(fd_t fd, int64_t offset, seek_origin_t origin);
+size_t seek(fd_t fd, ssize_t offset, seek_origin_t origin);
 
 /**
  * @brief System call for changing the cwd.
@@ -405,7 +406,7 @@ uint64_t stat(const char* path, stat_t* stat);
  * @return On success, the return value depends on the driver but is usually 0. On failure, `ERR` and `errno` is
  * set.
  */
-uint64_t ioctl(fd_t fd, uint64_t request, void* argp, uint64_t size);
+uint64_t ioctl(fd_t fd, uint64_t request, void* argp, size_t size);
 
 /**
  * @brief System call for duplicating file descriptors.
@@ -456,7 +457,7 @@ typedef struct
  * @return On success, the total number of bytes written to the buffer. On failure,
  * returns `ERR` and `errno` is set.
  */
-uint64_t getdents(fd_t fd, dirent_t* buffer, uint64_t count);
+size_t getdents(fd_t fd, dirent_t* buffer, uint64_t count);
 
 /**
  * @brief Helper for reading all directory entries.
@@ -468,7 +469,7 @@ uint64_t getdents(fd_t fd, dirent_t* buffer, uint64_t count);
  * @param count Output pointer to store the number of bytes written to the buffer.
  * @return On success, `0`. On failure, `ERR` and `errno` is set.
  */
-uint64_t readdir(fd_t fd, dirent_t** buffer, uint64_t* count);
+size_t readdir(fd_t fd, dirent_t** buffer, uint64_t* count);
 
 /**
  * @brief Wrapper for creating a directory.
@@ -581,7 +582,7 @@ uint64_t unmount(const char* mountpoint);
  * @param count The size of the buffer.
  * @return On success, the number of bytes read. On failure, `ERR` and `errno` is set.
  */
-uint64_t readlink(const char* path, char* buffer, uint64_t count);
+size_t readlink(const char* path, char* buffer, uint64_t count);
 
 /**
  * @brief System call for creating a symbolic link.
